Create coffe-type' crud. (#10)
* Create coffe-type' crud.

* Create AttachmentController, AttachmentService, AttachmentDto ProductController, ProductService and create the validators  for these functions.

* Fix issues for product crud flow.

+++ La-Mula-App-main/src/main/kotlin/com/lamulaapp/service/ProductService.kt
@@ -3,8 +3,12 @@ package com.lamulaapp.service
 import com.lamulaapp.controller.dto.ProductDto
 import com.lamulaapp.controller.mapper.toDto
 import com.lamulaapp.controller.mapper.toEntity
+import com.lamulaapp.exception.DuplicateKeyException
+import com.lamulaapp.exception.KeysAreDifferentException
 import com.lamulaapp.repository.ProductRepository
+import jakarta.persistence.EntityNotFoundException
 import org.springframework.stereotype.Service
+import java.time.LocalDateTime
 import java.util.*
 
 @Service
@@ -13,6 +17,11 @@ class ProductService(
 ) {
 
     fun createProduct(productDto: ProductDto): ProductDto {
+        val responsePKFound = productDto.idProduct?.let { productRepository.findById(it) }
+
+        if (responsePKFound != null && responsePKFound.isPresent) {
+            throw DuplicateKeyException("This ID already exists for the product to be created!")
+        }
         val response = productRepository.save(productDto.toEntity())
         return response.toDto()
     }
@@ -26,7 +35,7 @@ class ProductService(
         return if (response.isPresent) {
             response.get().toDto()
         } else {
-            null
+            throw EntityNotFoundException("There is no product with the given ID!")
         }
     }
 
@@ -34,16 +43,30 @@ class ProductService(
         val response = productRepository.findById(id)
 
         if (!response.isPresent) {
-            return null
+            throw EntityNotFoundException("There is no product with the given ID!")
         }
 
         if (id != productDto.idProduct) {
-            return null
+            throw KeysAreDifferentException("The keys should be the same as the product with the given ID!")
         }
-        return productRepository.save(productDto.toEntity()).toDto()
+
+        if (productDto.updatedBy == null) {
+            throw IllegalArgumentException("The field 'updatedBy' is mandatory!")
+        }
+
+        val entity = productDto
+            .toEntity()
+            .copy(updatedAt = LocalDateTime.now(), updatedBy = productDto.updatedBy)
+
+        return productRepository.save(entity).toDto()
     }
 
     fun deleteProduct(id: UUID) {
+       val response = productRepository.findById(id)
+
+        if (!response.isPresent) {
+            throw EntityNotFoundException("There is no product with the given ID!")
+        }
         productRepository.deleteById(id)
     }
 }
